debug

+++ crates/aptos-dkg/src/dlog/bsgs.rs
@@ -10,16 +10,16 @@ use std::collections::HashMap;
 /// - `G`: base of the exponentiation
 /// - `H`: target point
 /// - `baby_table`: precomputed HashMap from `C.to_compressed()` |---> exponent
-/// - `range_limit`: maximum size of the exponent we're trying to obtain. TODO: Change to u64?
+/// - `range_limit`: maximum size of the exponent we're trying to obtain
 //
 // TODO:: ensure that G is also the element used to build the baby_table? So turn baby_table into a struct?
 #[allow(non_snake_case)]
 pub fn dlog<C: CurveGroup>(
     G: C,
     H: C,
-    baby_table: &HashMap<Vec<u8>, u32>,
-    range_limit: u32,
-) -> Option<u32> {
+    baby_table: &HashMap<Vec<u8>, u64>,
+    range_limit: u64,
+) -> Option<u64> {
     let byte_size = G.compressed_size();
 
     let m = baby_table
@@ -50,9 +50,9 @@ pub fn dlog<C: CurveGroup>(
 pub fn dlog_vec<C: CurveGroup>(
     G: C,
     H_vec: &[C],
-    baby_table: &HashMap<Vec<u8>, u32>,
-    range_limit: u32,
-) -> Option<Vec<u32>> {
+    baby_table: &HashMap<Vec<u8>, u64>,
+    range_limit: u64,
+) -> Option<Vec<u64>> {
     let mut result = Vec::with_capacity(H_vec.len());
 
     for H in H_vec {
